Work on setting correct size of uploaded picture.
