+++ photo-wall-planner/Assets/photo-wall-planner/assets/Frames/20x25 LS.prefab
@@ -8,6 +8,12 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 479698811589167714, guid: 14f6e80527ed0014d80295e2e18a1e1f,
+        type: 3}
+      propertyPath: m_Sprite
+      value: 
+      objectReference: {fileID: 5706957933644387078, guid: 215687feeab0b7445b4a1cd86c82c2e1,
+        type: 3}
     - target: {fileID: 2644699653642223145, guid: 14f6e80527ed0014d80295e2e18a1e1f,
         type: 3}
       propertyPath: m_Name
@@ -63,7 +69,8 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 4432460469505752554, guid: 14f6e80527ed0014d80295e2e18a1e1f, type: 3}
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
     m_AddedComponents: []
