+++ photo-wall-planner/Assets/photo-wall-planner/scripts/GalleryManager.cs.meta
